% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.intensity.R
\name{filter.intensity}
\alias{filter.intensity}
\title{Filter a time-series vector/matrix based on its spectral intensity}
\usage{
filter.intensity(x, centered = TRUE, centred = centered, scaled = TRUE,
         highpass = NULL, highpass.remove = TRUE,
         lowpass  = NULL, lowpass.remove  = TRUE)
}
\arguments{
\item{x}{A vector of time-series values}

\item{centered}{Logical; if \code{TRUE} the time-series vector is centred (mean removed) before computing its intensity}

\item{centred}{An alternative name for the \code{centered} input}

\item{scaled}{Logical; if \code{TRUE} the intensity measure is scaled so that its norm is equal to the number of values in the time-series}

\item{highpass}{The minimum intensity value for the highpass filter (or NULL if there is no highpass filter)}

\item{highpass.remove}{Logical; if \code{TRUE} then the highpass filter removes (instead of attenuating) the low-frequency signal}

\item{lowpass}{The maximum intensity value for the lowpass filter (or NULL if there is no lowpass filter)}

\item{lowpass.remove}{Logical; if \code{TRUE} then the lowpass filter removes (instead of attenuating) the high-frequency signal}
}
\description{
\code{filter.intensity} returns a vector/matrix of filtered time-series
}
\details{
This function filters a time-series based on its spectral intensity.  The user inputs the time-series vector or matrix and
specifies whether it is to be centered and/or scaled.  Centering subtracts the sample mean of the vector prior to conversion
into frequency space; this sets the intensity of the signal to zero at the zero frequency.  Scaling scales the intensity so
the that norm of the intensity vector is equal to the number of values in the time-series.  The function then applies the
specified highpass/lowpass filter and returns the filtered signals.  By default, both the highpass and lowpass filters remove
parts of the signal that do not meet the filter requirements; if preferred the filters can attenuate (rather than removing)
these signals.
}
