% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{garma}
\alias{garma}
\alias{SERIES}
\alias{SERIES1}
\title{Simulated example data set}
\description{
A simulated set of 16 time series over 30 time points.
}
\details{
Created using \code{set.seed(1014745590)}
}
