% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spectrum.test.R
\name{plot.spectrum.test}
\alias{plot.spectrum.test}
\title{Plot of the Permutation-Spectrum Test}
\usage{
\method{plot}{spectrum.test}(test, ggplot = TRUE, print = TRUE, ...)
}
\arguments{
\item{test}{A ```spectrum.test``` object produced by the ```spectrum.test``` function}

\item{ggplot}{Logical; if ```TRUE``` the scatterplot is a ```ggplot``` object; if ```FALSE``` it is a ```base``` plot object}

\item{print}{Logical; if ```TRUE``` the scatterplot is printed}
}
\description{
This function generates a dual plot showing the results of the permutation-spectrum testusing either ```ggplot``` or ```base``` graphics.
The user must input the time-series vector ```x``` and either a vector of simulated values of the maximum scaled intensity, or an entire
test object produced by the ```spectrum.test``` object.  The function produces a dual plot showing the scaled intensity of the time-series
vector and the simulated null distribution of the maximum scaled intensity under the null hypothesis of an IID vector.  The plots also
report the value of the maximum scaled intensity and the resulting p-value for the test.  This dual plot forms a useful companion to the
permutation-spectrum test; it allows the user to visualise the simulated null distribution and test statistic.
}
